Rotation sample ported to 64-bits

+++ Rotation/HandlingRotation/Screens/iPad/Method1Autosize/AutosizeScreenPad.xib.cs
@@ -2,8 +2,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 
 namespace HandlingRotation.Screens.iPad.Method1Autosize {
 	public partial class AutosizeScreenPad : UIViewController {
@@ -104,7 +104,7 @@ namespace HandlingRotation.Screens.iPad.Method1Autosize {
 			}
 			
 			// set the cell properties
-			cell.TextLabel.Text = this.tableItems[indexPath.Row];
+			cell.TextLabel.Text = this.tableItems[(int)indexPath.Row];
 			
 			// return the cell
 			return cell;
